Crate mozjpeg_sys

Please read libjpeg manual for description of functions in this crate.

Re-exports

pub use J_DCT_METHOD::JDCT_IFAST as JDCT_FASTEST;

pub use J_DCT_METHOD::JDCT_ISLOW as JDCT_DEFAULT;

pub use JINT_COMPRESS_PROFILE_VALUE::*;

pub use J_BOOLEAN_PARAM::*;

pub use J_COLOR_SPACE::*;

pub use J_FLOAT_PARAM::*;

pub use J_INT_PARAM::*;

Structs

JHUFF_TBL

JQUANT_TBL

jpeg_common_struct

Routines that are to be used by both halves of the library are declared to receive a pointer to this structure. There are no actual instances of jpeg_common_struct, only of jpeg_compress_struct and jpeg_decompress_struct.

jpeg_component_info

jpeg_compress_struct

jpeg_decompress_struct

jpeg_destination_mgr

jpeg_error_mgr

Error handler object

jpeg_marker_struct

jpeg_memory_mgr

jpeg_progress_mgr

jpeg_scan_info

jpeg_source_mgr

jvirt_barray_control

This is an opaque type. Don’t assume size or alignment of this struct.

jvirt_sarray_control

This is an opaque type. Don’t assume size or alignment of this struct.

msg_parm_union

Enums

FILE

JINT_COMPRESS_PROFILE_VALUE

J_BOOLEAN_PARAM

These 32-bit GUIDs and the corresponding jpeg_*_get_*_param() jpeg_*_set_*_param() functions allow for extending the libjpeg API without breaking backward ABI compatibility. The actual parameters are stored in the opaque jpeg_comp_master and jpeg_decomp_master structs.

J_COLOR_SPACE

J_DCT_METHOD

J_DITHER_MODE

J_FLOAT_PARAM

J_INT_PARAM

jpeg_marker

Constants

DCTSIZE

The basic DCT block is 8x8 samples

DCTSIZE2

DCTSIZE²

JERR_ARITH_NOTIMPL

JERR_BAD_ALIGN_TYPE

JERR_BAD_ALLOC_CHUNK

JERR_BAD_BUFFER_MODE

JERR_BAD_COMPONENT_ID

JERR_BAD_CROP_SPEC

JERR_BAD_DCTSIZE

JERR_BAD_DCT_COEF

JERR_BAD_DROP_SAMPLING

JERR_BAD_HUFF_TABLE

JERR_BAD_IN_COLORSPACE

JERR_BAD_J_COLORSPACE

JERR_BAD_LENGTH

JERR_BAD_LIB_VERSION

JERR_BAD_MCU_SIZE

JERR_BAD_PARAM

JERR_BAD_PARAM_VALUE

JERR_BAD_POOL_ID

JERR_BAD_PRECISION

JERR_BAD_PROGRESSION

JERR_BAD_PROG_SCRIPT

JERR_BAD_SAMPLING

JERR_BAD_SCAN_SCRIPT

JERR_BAD_STATE

JERR_BAD_STRUCT_SIZE

JERR_BAD_VIRTUAL_ACCESS

JERR_BUFFER_SIZE

JERR_CANT_SUSPEND

JERR_CCIR601_NOTIMPL

JERR_COMPONENT_COUNT

JERR_CONVERSION_NOTIMPL

JERR_DAC_INDEX

JERR_DAC_VALUE

JERR_DHT_INDEX

JERR_DQT_INDEX

JERR_EMPTY_IMAGE

JERR_EMS_READ

JERR_EMS_WRITE

JERR_EOI_EXPECTED

JERR_FILE_READ

JERR_FILE_WRITE

JERR_FRACT_SAMPLE_NOTIMPL

JERR_HUFF_CLEN_OVERFLOW

JERR_HUFF_MISSING_CODE

JERR_IMAGE_TOO_BIG

JERR_INPUT_EMPTY

JERR_INPUT_EOF

JERR_MISMATCHED_QUANT_TABLE

JERR_MISSING_DATA

JERR_MODE_CHANGE

JERR_NOTIMPL

JERR_NOT_COMPILED

JERR_NO_BACKING_STORE

JERR_NO_HUFF_TABLE

JERR_NO_IMAGE

JERR_NO_QUANT_TABLE

JERR_NO_SOI

JERR_OUT_OF_MEMORY

JERR_QUANT_COMPONENTS

JERR_QUANT_FEW_COLORS

JERR_QUANT_MANY_COLORS

JERR_SOF_DUPLICATE

JERR_SOF_NO_SOS

JERR_SOF_UNSUPPORTED

JERR_SOI_DUPLICATE

JERR_SOS_NO_SOF

JERR_TFILE_CREATE

JERR_TFILE_READ

JERR_TFILE_SEEK

JERR_TFILE_WRITE

JERR_TOO_LITTLE_DATA

JERR_UNKNOWN_MARKER

JERR_UNSUPPORTED_SUSPEND

JERR_VIRTUAL_BUG

JERR_WIDTH_OVERFLOW

JERR_XMS_READ

JERR_XMS_WRITE

JMSG_COPYRIGHT

JMSG_LASTMSGCODE

JMSG_NOMESSAGE

JMSG_VERSION

JPEG_LIB_VERSION

JPOOL_IMAGE

lasts until done with image/datastream

JPOOL_PERMANENT

lasts until master record is destroyed

JTRC_16BIT_TABLES

JTRC_ADOBE

JTRC_APP0

JTRC_APP14

JTRC_DAC

JTRC_DHT

JTRC_DQT

JTRC_DRI

JTRC_EMS_CLOSE

JTRC_EMS_OPEN

JTRC_EOI

JTRC_HUFFBITS

JTRC_JFIF

JTRC_JFIF_BADTHUMBNAILSIZE

JTRC_JFIF_EXTENSION

JTRC_JFIF_THUMBNAIL

JTRC_MISC_MARKER

JTRC_PARMLESS_MARKER

JTRC_QUANTVALS

JTRC_QUANT_3_NCOLORS

JTRC_QUANT_NCOLORS

JTRC_QUANT_SELECTED

JTRC_RECOVERY_ACTION

JTRC_RST

JTRC_SMOOTH_NOTIMPL

JTRC_SOF

JTRC_SOF_COMPONENT

JTRC_SOI

JTRC_SOS

JTRC_SOS_COMPONENT

JTRC_SOS_PARAMS

JTRC_TFILE_CLOSE

JTRC_TFILE_OPEN

JTRC_THUMB_JPEG

JTRC_THUMB_PALETTE

JTRC_THUMB_RGB

JTRC_UNKNOWN_IDS

JTRC_XMS_CLOSE

JTRC_XMS_OPEN

JWRN_ADOBE_XFORM

JWRN_BOGUS_ICC

JWRN_BOGUS_PROGRESSION

JWRN_EXTRANEOUS_DATA

JWRN_HIT_MARKER

JWRN_HUFF_BAD_CODE

JWRN_JFIF_MAJOR

JWRN_JPEG_EOF

JWRN_MUST_RESYNC

JWRN_NOT_SEQUENTIAL

JWRN_TOO_MUCH_DATA

NUM_QUANT_TBLS

Quantization tables are numbered 0..3

Functions

jpeg_CreateCompress⚠

jpeg_CreateDecompress⚠

jpeg_abort⚠

jpeg_abort_compress⚠

jpeg_abort_decompress⚠

jpeg_add_quant_table⚠

jpeg_alloc_huff_table⚠

jpeg_alloc_quant_table⚠

jpeg_c_bool_param_supported⚠

jpeg_c_float_param_supported⚠

jpeg_c_get_bool_param⚠

jpeg_c_get_float_param⚠

jpeg_c_get_int_param⚠

jpeg_c_int_param_supported⚠

jpeg_c_set_bool_param⚠

jpeg_c_set_float_param⚠

jpeg_c_set_int_param⚠

jpeg_calc_output_dimensions⚠

jpeg_consume_input⚠

jpeg_copy_critical_parameters⚠

jpeg_create_compress⚠

jpeg_create_decompress⚠

jpeg_crop_scanline⚠

This function provides application programmers with the ability to decompress only a portion of each row in the JPEG image. It must be called after jpeg_start_decompress() and before any calls to jpeg_read_scanlines() or jpeg_skip_scanlines(). If the output image is scaled, then xoffset and width are relative to the scaled image dimensions. xoffset and width are passed by reference because xoffset must fall on an iMCU boundary. If it doesn’t, then it will be moved left to the nearest iMCU boundary, and width will be increased accordingly.

jpeg_default_colorspace⚠

jpeg_destroy⚠

jpeg_destroy_compress⚠

jpeg_destroy_decompress⚠

jpeg_finish_compress⚠

jpeg_finish_decompress⚠

jpeg_finish_output⚠

jpeg_float_quality_scaling⚠

jpeg_has_multiple_scans⚠

jpeg_input_complete⚠

jpeg_mem_dest⚠

jpeg_mem_src⚠

jpeg_new_colormap⚠Deprecated

jpeg_quality_scaling⚠

jpeg_read_coefficients⚠

jpeg_read_header⚠

jpeg_read_raw_data⚠

jpeg_read_scanlines⚠

jpeg_resync_to_restart⚠

jpeg_save_markers⚠

jpeg_set_colorspace⚠

jpeg_set_defaults⚠

jpeg_set_idct_method_selector⚠

jpeg_set_linear_quality⚠

jpeg_set_marker_processor⚠

jpeg_set_quality⚠

jpeg_simple_progression⚠

jpeg_skip_scanlines⚠

This function provides application programmers with the ability to skip over multiple rows in the JPEG image.

jpeg_start_compress⚠

jpeg_start_decompress⚠

jpeg_start_output⚠

jpeg_std_error⚠

jpeg_stdio_dest⚠

jpeg_stdio_src⚠

jpeg_suppress_tables⚠

jpeg_write_coefficients⚠

jpeg_write_m_byte⚠

jpeg_write_m_header⚠

jpeg_write_marker⚠

jpeg_write_raw_data⚠

jpeg_write_scanlines⚠

jpeg_write_tables⚠

Type Aliases

DCTELEM

type for individual integer DCT coefficient

JBLOCK

one block of coefficients

JBLOCKARRAY

JBLOCKROW

pointer to one row of coefficient blocks

JCOEF

JDIMENSION

JSAMPARRAY

ptr to some rows (a 2-D sample array)

JSAMPARRAY_MUT

JSAMPIMAGE

a 3-D sample array: top index is color

JSAMPIMAGE_MUT

JSAMPLE

JSAMPROW

ptr to one image row of pixel samples.

JSAMPROW_MUT

J_MESSAGE_CODE

boolean

c_int

Equivalent to C’s signed int (int) type.

c_long

Equivalent to C’s signed long (long) type.

c_uint

Equivalent to C’s unsigned int type.

c_ulong

Equivalent to C’s unsigned long type.

c_void

Equivalent to C’s void type when used as a pointer.

jpeg_marker_parser_method