mozjpeg_sys

Struct jpeg_compress_struct

pub struct jpeg_compress_struct {

    pub common: jpeg_common_struct,
    pub dest: *mut jpeg_destination_mgr,
    pub image_width: JDIMENSION,
    pub image_height: JDIMENSION,
    pub input_components: c_int,
    pub in_color_space: J_COLOR_SPACE,
    pub input_gamma: f64,
    pub data_precision: c_int,
    pub num_components: c_int,
    pub jpeg_color_space: J_COLOR_SPACE,
    pub comp_info: *mut jpeg_component_info,
    pub quant_tbl_ptrs: [*mut JQUANT_TBL; 4],
    pub dc_huff_tbl_ptrs: [*mut JHUFF_TBL; 4],
    pub ac_huff_tbl_ptrs: [*mut JHUFF_TBL; 4],
    pub arith_dc_L: [u8; 16],
    pub arith_dc_U: [u8; 16],
    pub arith_ac_K: [u8; 16],
    pub num_scans: c_int,
    pub scan_info: *const jpeg_scan_info,
    pub raw_data_in: boolean,
    pub arith_code: boolean,
    pub optimize_coding: boolean,
    pub CCIR601_sampling: boolean,
    pub smoothing_factor: c_int,
    pub dct_method: J_DCT_METHOD,
    pub restart_interval: c_uint,
    pub restart_in_rows: c_int,
    pub write_JFIF_header: boolean,
    pub JFIF_major_version: u8,
    pub JFIF_minor_version: u8,
    pub density_unit: u8,
    pub X_density: u16,
    pub Y_density: u16,
    pub write_Adobe_marker: boolean,
    pub next_scanline: JDIMENSION,
    pub max_h_samp_factor: c_int,
    pub max_v_samp_factor: c_int,
    /* private fields */

}

Fields
common: jpeg_common_structdest: *mut jpeg_destination_mgrimage_width: JDIMENSION

Description of source image — these fields must be filled in by outer application before starting compression.

image_height: JDIMENSIONinput_components: c_intin_color_space: J_COLOR_SPACE

in_color_space must be correct before you can even call jpeg_set_defaults().

input_gamma: f64

image gamma of input image

data_precision: c_int

bits of precision in image data

num_components: c_intjpeg_color_space: J_COLOR_SPACEcomp_info: *mut jpeg_component_info

comp_info[i] describes component that appears i’th in SOF

quant_tbl_ptrs: [*mut JQUANT_TBL; 4]dc_huff_tbl_ptrs: [*mut JHUFF_TBL; 4]

ptrs to coefficient quantization tables, or NULL if not defined, and corresponding scale factors (percentage, initialized 100).

ac_huff_tbl_ptrs: [*mut JHUFF_TBL; 4]arith_dc_L: [u8; 16]arith_dc_U: [u8; 16]arith_ac_K: [u8; 16]num_scans: c_intscan_info: *const jpeg_scan_inforaw_data_in: boolean

TRUE=caller supplies downsampled data

arith_code: booleanoptimize_coding: boolean

TRUE=optimize entropy encoding parms

CCIR601_sampling: booleansmoothing_factor: c_intdct_method: J_DCT_METHODrestart_interval: c_uint

MCUs per restart, or 0 for no restart

restart_in_rows: c_intwrite_JFIF_header: booleanJFIF_major_version: u8JFIF_minor_version: u8density_unit: u8X_density: u16Y_density: u16write_Adobe_marker: booleannext_scanline: JDIMENSIONmax_h_samp_factor: c_intmax_v_samp_factor: c_int