mozjpeg_sys

Struct jpeg_decompress_struct

pub struct jpeg_decompress_struct {

    pub common: jpeg_common_struct,
    pub src: *mut jpeg_source_mgr,
    pub image_width: JDIMENSION,
    pub image_height: JDIMENSION,
    pub num_components: c_int,
    pub jpeg_color_space: J_COLOR_SPACE,
    pub out_color_space: J_COLOR_SPACE,
    pub scale_num: c_uint,
    pub scale_denom: c_uint,
    pub output_gamma: f64,
    pub buffered_image: boolean,
    pub raw_data_out: boolean,
    pub dct_method: J_DCT_METHOD,
    pub do_fancy_upsampling: boolean,
    pub do_block_smoothing: boolean,
    pub quantize_colors: boolean,
    pub dither_mode: J_DITHER_MODE,
    pub two_pass_quantize: boolean,
    pub desired_number_of_colors: c_int,
    pub enable_1pass_quant: boolean,
    pub enable_external_quant: boolean,
    pub enable_2pass_quant: boolean,
    pub output_width: JDIMENSION,
    pub output_height: JDIMENSION,
    pub out_color_components: c_int,
    pub output_components: c_int,
    pub rec_outbuf_height: c_int,
    pub actual_number_of_colors: c_int,
    pub colormap: JSAMPARRAY_MUT,
    pub output_scanline: JDIMENSION,
    pub input_scan_number: c_int,
    pub input_iMCU_row: JDIMENSION,
    pub output_scan_number: c_int,
    pub output_iMCU_row: JDIMENSION,
    pub coef_bits: *mut c_void,
    pub quant_tbl_ptrs: [*mut JQUANT_TBL; 4],
    pub comp_info: *mut jpeg_component_info,
    pub density_unit: u8,
    pub X_density: u16,
    pub Y_density: u16,
    pub marker_list: *mut jpeg_marker_struct,
    pub max_h_samp_factor: c_int,
    pub max_v_samp_factor: c_int,
    /* private fields */

}

Fields
common: jpeg_common_structsrc: *mut jpeg_source_mgrimage_width: JDIMENSION

Basic description of image — filled in by jpeg_read_header()

image_height: JDIMENSIONnum_components: c_intjpeg_color_space: J_COLOR_SPACEout_color_space: J_COLOR_SPACE

Decompression processing parameters — these fields must be set before calling jpeg_start_decompress(). Note that jpeg_read_header() initializes them to default values.

scale_num: c_uintscale_denom: c_uintoutput_gamma: f64

image gamma wanted in output

buffered_image: booleanraw_data_out: boolean

TRUE=downsampled data wanted

dct_method: J_DCT_METHODdo_fancy_upsampling: booleando_block_smoothing: booleanquantize_colors: booleandither_mode: J_DITHER_MODEtwo_pass_quantize: booleandesired_number_of_colors: c_intenable_1pass_quant: booleanenable_external_quant: booleanenable_2pass_quant: booleanoutput_width: JDIMENSION

Description of actual output image that will be returned to application. These fields are computed by jpeg_start_decompress(). You can also use jpeg_calc_output_dimensions() to determine these values in advance of calling jpeg_start_decompress().

output_height: JDIMENSIONout_color_components: c_intoutput_components: c_intrec_outbuf_height: c_int

min recommended height of scanline buffer If the buffer passed to jpeg_read_scanlines() is less than this many rows high, space and time will be wasted due to unnecessary data copying. Usually rec_outbuf_height will be 1 or 2, at most 4.

actual_number_of_colors: c_intcolormap: JSAMPARRAY_MUToutput_scanline: JDIMENSION

Row index of next scanline to be read from jpeg_read_scanlines(). Application may use this to control its processing loop, e.g., “while (output_scanline < output_height)”.

input_scan_number: c_int

Current input scan number and number of iMCU rows completed in scan. These indicate the progress of the decompressor input side.

input_iMCU_row: JDIMENSIONoutput_scan_number: c_intoutput_iMCU_row: JDIMENSIONcoef_bits: *mut c_void

Current progression status. coef_bits[c][i] indicates the precision with which component c’s DCT coefficient i (in zigzag order) is known. It is -1 when no data has yet been received, otherwise it is the point transform (shift) value for the most recent scan of the coefficient (thus, 0 at completion of the progression). This pointer is NULL when reading a non-progressive file.

quant_tbl_ptrs: [*mut JQUANT_TBL; 4]

Internal JPEG parameters — the application usually need not look at these fields. Note that the decompressor output side may not use any parameters that can change between scans.

comp_info: *mut jpeg_component_infodensity_unit: u8X_density: u16Y_density: u16marker_list: *mut jpeg_marker_structmax_h_samp_factor: c_int

These fields are computed during decompression startup

max_v_samp_factor: c_int